// services/document-evidence/Generated/frontend/documentfragments/types/CreateDocumentFragmentRequest.generated.ts
// ------------------------------------------------------------------------------
// GENERATED FILE - source: documentfragments/types.json
// DO NOT EDIT MANUALLY.
// ------------------------------------------------------------------------------

export interface CreateDocumentFragmentRequest {
  "anchorJson": string;
  "contentHash": string;
  "documentRevisionId": string;
  "kind": string;
  "sequence": number;
  "text": string;
}
export const CreateDocumentFragmentRequestTypeKey = "LogicLens.DocumentEvidence.Generated.Api.Contracts.CreateDocumentFragmentRequest" as const;